docs: Make the demo code reusable by spliting it into pieces
