+++ docz/Demo/Demo.js
@@ -1,40 +1,11 @@
 import React from 'react';
 import { Html5Table as Table } from '../../';
-import data from './data';
-import Bootstrap from '../bootstrap';
+import { data, columns } from './helpers';
 import './demo.scss';
-
-const Avatar = ({ row, column }) => {
-  return (
-    <img
-      src={row[column.key]}
-      alt="avatar"
-      style={{
-        height: '40px'
-      }}
-    />
-  );
-};
-
-const IndexCount = ({ index }) => {
-  return <div>{index}</div>;
-};
-
-const columns = [
-  { key: 'index', width: 25, title: 'Index', Component: IndexCount },
-  { key: 'avatar', width: 40, title: 'Avatar', Component: Avatar },
-  { key: 'name', width: 100, title: 'Name' },
-  { key: 'clan', width: 100, title: 'Clan' },
-  { key: 'age', width: 40, title: 'Age' }
-];
 
 function ShinobiTable(props) {
   return (
-    <Bootstrap
-      style={{
-        height: '250px'
-      }}
-    >
+    <div className="bootstrap-wrapper" style={{ height: '250px' }}>
       <Table
         data={data}
         columns={columns}
@@ -42,7 +13,7 @@ function ShinobiTable(props) {
         headerClassName="thead-dark"
         {...props}
       />
-    </Bootstrap>
+    </div>
   );
 }
